refactor cw encoder
